+++ src/ts/beatmap/contributor.ts
@@ -2,4 +2,5 @@ export interface Contributor {
     _role: string;
     _name: string;
     _iconPath: string;
+    _base64?: string;
 }