+++ src/ts/beatmap/customData.ts
@@ -1,8 +1,16 @@
-import { ChromaEnvironmentOld, ChromaEvent, ChromaNote, ChromaObstacle } from './chroma';
+import { Bookmark } from './bookmark';
+import { BPMChange } from './bpm';
+import {
+    ChromaEnvironment,
+    ChromaEnvironmentOld,
+    ChromaEvent,
+    ChromaNote,
+    ChromaObstacle,
+} from './chroma';
 import { Contributor } from './contributor';
 import { Editor } from './editor';
 import { ColorScheme } from './environment';
-import { NEEvent, NENote, NEObstacle } from './noodleExtensions';
+import { NECustomData, NEEvent, NENote, NEObstacle } from './noodleExtensions';
 
 export interface CustomData {
     [key: string]: any;
@@ -23,6 +31,13 @@ export interface CustomDataInfoDifficulty extends CustomData, ColorScheme, Chrom
     _information?: string[];
     _suggestions?: string[];
     _requirements?: string[];
+}
+
+export interface CustomDataDifficulty extends CustomData, ChromaEnvironment, NECustomData {
+    _time?: number;
+    _bpmChanges?: BPMChange[];
+    _BPMChanges?: BPMChange[];
+    _bookmarks?: Bookmark[];
 }
 
 export interface CustomDataNote extends CustomData, ChromaNote, NENote {}