+++ src/ts/beatmap/difficulty.ts
@@ -1,5 +1,13 @@
 export type DifficultyName = 'Easy' | 'Normal' | 'Hard' | 'Expert' | 'ExpertPlus';
 
+export enum DifficultyRename {
+    'Easy' = 'Easy',
+    'Normal' = 'Normal',
+    'Hard' = 'Hard',
+    'Expert' = 'Expert',
+    'ExpertPlus' = 'Expert+',
+}
+
 export enum DifficultyRank {
     'Easy' = 1,
     'Normal' = 3,