+++ src/ts/tools/analyse.ts
@@ -1,0 +1,11 @@
+import savedData from '../savedData';
+import { CharacteristicName } from '../beatmap/characteristic';
+import { DifficultyName } from '../beatmap/difficulty';
+
+const general = (): void => {};
+const difficulty = (mode?: CharacteristicName, difficulty?: DifficultyName): void => {
+    if (mode && difficulty) {
+    }
+};
+
+export default { general, difficulty };