+++ src/ts/ui/settings.ts
@@ -4,31 +4,31 @@ const htmlBody = document.querySelector<HTMLBodyElement>('body');
 const htmlSettingsTheme = document.querySelectorAll<HTMLSelectElement>('.settings__theme');
 const htmlSettingsShow = document.querySelectorAll<HTMLInputElement>('.settings__show');
 
-htmlSettingsTheme.forEach((elem) => elem.addEventListener('change', settingsThemeHandler));
-htmlSettingsShow.forEach((elem) => elem.addEventListener('click', settingsShowHandler));
+htmlSettingsTheme.forEach((elem) => elem.addEventListener('change', themeChangeHandler));
+htmlSettingsShow.forEach((elem) => elem.addEventListener('click', showCheckHandler));
 
-function settingsThemeHandler(ev: Event): void {
+function themeChangeHandler(ev: Event): void {
     const target = ev.target as HTMLSelectElement;
     settings.theme = target.options[target.options.selectedIndex].value as Theme;
     htmlBody!.className = 'theme-' + settings.theme.toLowerCase().replace(' ', '');
     settings.save();
 }
 
-function settingsShowHandler(ev: Event): void {
+function showCheckHandler(ev: Event): void {
     const target = ev.target as HTMLInputElement;
     const id = target.id.replace('settings__show-', '');
     settings.show[id] = target.checked;
     settings.save();
 }
 
-export const settingsSetTheme = (str: Theme): void => {
+export const setTheme = (str: Theme): void => {
     htmlBody!.className = 'theme-' + str.toLowerCase().replace(' ', '');
     htmlSettingsTheme.forEach((elem) => {
         elem.value = str;
     });
 };
 
-export const settingsShowSetCheck = (id: string, bool: boolean): void => {
+export const setShowCheck = (id: string, bool: boolean): void => {
     htmlSettingsShow.forEach((elem) => {
         if (elem.id.endsWith(id)) {
             elem.checked = bool;
@@ -36,10 +36,14 @@ export const settingsShowSetCheck = (id: string, bool: boolean): void => {
     });
 };
 
-export const settingsClear = (): void => {
+export const clear = (): void => {
     settings.clear();
     settings.reset();
-    for (const key in settings.load) {
-        settingsShowSetCheck(key, settings.load[key]);
-    }
+    location.reload(true);
+};
+
+export default {
+    setTheme,
+    setShowCheck,
+    clear,
 };