+++ src/ts/ui/stats.ts
@@ -1,0 +1,8 @@
+import { CharacteristicName } from '../beatmap/characteristic';
+import { BeatmapInfo, BeatmapSetDifficulty } from '../beatmap/info';
+import { BeatmapData } from '../beatmap/map';
+
+export const populateMode = (): void => {};
+export const populateDiff = (): void => {};
+
+export default { populateMode, populateDiff };