+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## v2.0.0 : TypeScript Rewrite
 
+### Added
+
+-   Information Tab
+-   -   SongCore, Chroma & Noodle Extensions related stuff
+-   Stats Tab
+-   Settings Tab
+-   Theme
+
 ### Changed
 
 -   Codebase rewritten